+++ scalafix-cli/src/main/scala/scalafix/cli/WriteMode.scala
@@ -1,0 +1,10 @@
+package scalafix.cli
+
+sealed abstract class WriteMode {
+  def isWriteFile: Boolean = this == WriteMode.WriteFile
+}
+
+object WriteMode {
+  case object WriteFile extends WriteMode
+  case object Stdout extends WriteMode
+}
