+++ scalafix-tests/src/test/resources/checkSyntax/NoAutoTupling.source
@@ -1,4 +1,4 @@
-ONLY rewrites = [NoAutoTupling]
+rewrites = [NoAutoTupling]
 <<< add explicit tuple
 object tup {
   def fooo(t: (Int, Int)): Int = ???
